Reorganize WG kernel tunnel
