+++ talpid-core/src/tunnel/wireguard/wireguard_kernel/nm_tunnel.rs
@@ -1,4 +1,7 @@
-use super::{config::Config, stats::Stats, Tunnel, TunnelError};
+use super::{
+    super::stats::Stats, wg_message::DeviceNla, Config, Error as WgKernelError, Handle, Tunnel,
+    TunnelError, MULLVAD_INTERFACE_NAME,
+};
 use dbus::{
     arg::{RefArg, Variant},
     blocking::{stdintf::org_freedesktop_dbus::Properties, BlockingSender, Connection, Proxy},
@@ -53,21 +56,7 @@ pub enum Error {
     FindInterfaceIndex,
 }
 
-impl Error {
-    pub fn should_use_userspace(&self) -> bool {
-        match self {
-            Error::NMTooOld(_) => true,
-            _ => false,
-        }
-    }
-}
-
-use super::wireguard_kernel::{
-    wg_message::{DeviceNla, PeerNla},
-    Handle,
-};
-
-pub struct NetworkManager {
+pub struct NetworkManagerTunnel {
     dbus_connection: Connection,
     tunnel: Option<WireguardTunnel>,
     interface_index: u32,
@@ -81,21 +70,26 @@ type VariantMap = HashMap<String, VariantRefArg>;
 // settings are a{sa{sv}}
 type DbusSettings = HashMap<String, VariantMap>;
 
-impl NetworkManager {
-    pub fn new(tokio_handle: tokio::runtime::Handle, config: &Config) -> Result<Self> {
-        let mut dbus_connection = Connection::new_system().map_err(Error::Dbus)?;
-        Self::ensure_nm_is_new_enough(&dbus_connection)?;
-        let tunnel = Some(Self::create_wg_tunnel(&mut dbus_connection, config)?);
+impl NetworkManagerTunnel {
+    pub fn new(
+        tokio_handle: tokio::runtime::Handle,
+        config: &Config,
+    ) -> std::result::Result<Self, WgKernelError> {
+        let mut dbus_connection = Connection::new_system()
+            .map_err(|error| WgKernelError::NetworkManager(Error::Dbus(error)))?;
+        Self::ensure_nm_is_new_enough(&dbus_connection).map_err(WgKernelError::NetworkManager)?;
+        let tunnel = Some(
+            Self::create_wg_tunnel(&mut dbus_connection, config)
+                .map_err(WgKernelError::NetworkManager)?,
+        );
 
         tokio_handle.clone().block_on(async move {
-            let netlink_connections = Handle::connect()
-                .await
-                .map_err(|_| Error::FindInterfaceIndex)?;
+            let netlink_connections = Handle::connect().await?;
             let interface_index = Self::find_device_index(&netlink_connections)
                 .await?
-                .ok_or(Error::FindInterfaceIndex)?;
+                .ok_or(WgKernelError::NetworkManager(Error::FindInterfaceIndex))?;
 
-            Ok(NetworkManager {
+            Ok(NetworkManagerTunnel {
                 dbus_connection,
                 tunnel,
                 interface_index,
@@ -105,15 +99,11 @@ impl NetworkManager {
         })
     }
 
-    async fn find_device_index(netlink_connections: &Handle) -> Result<Option<u32>> {
+    async fn find_device_index(
+        netlink_connections: &Handle,
+    ) -> std::result::Result<Option<u32>, WgKernelError> {
         let mut wg = netlink_connections.wg_handle.clone();
-        let device = wg
-            .get_by_name("wg-mullvad".to_string())
-            .await
-            .map_err(|err| {
-                log::error!("Failed to fetch WireGuard device config: {}", err);
-                Error::FindInterfaceIndex
-            })?;
+        let device = wg.get_by_name(MULLVAD_INTERFACE_NAME.to_string()).await?;
 
         for nla in &device.nlas {
             if let DeviceNla::IfIndex(index) = nla {
@@ -269,10 +259,13 @@ impl NetworkManager {
         wireguard_config.insert("peers".into(), Variant(Box::new(peer_configs)));
 
         connection_config.insert("type".into(), Variant(Box::new("wireguard".to_string())));
-        connection_config.insert("id".into(), Variant(Box::new("wg-mullvad".to_string())));
+        connection_config.insert(
+            "id".into(),
+            Variant(Box::new(MULLVAD_INTERFACE_NAME.to_string())),
+        );
         connection_config.insert(
             "interface-name".into(),
-            Variant(Box::new("wg-mullvad".to_string())),
+            Variant(Box::new(MULLVAD_INTERFACE_NAME.to_string())),
         );
         connection_config.insert("autoconnect".into(), Variant(Box::new(true)));
 
@@ -359,7 +352,7 @@ impl NetworkManager {
     }
 }
 
-impl Tunnel for NetworkManager {
+impl Tunnel for NetworkManagerTunnel {
     fn get_interface_name(&self) -> String {
         if let Some(tunnel) = self.tunnel.as_ref() {
             let interface_name = tunnel
@@ -373,7 +366,7 @@ impl Tunnel for NetworkManager {
                 Err(error) => log::error!("Failed to fetch interface name from NM: {}", error),
             }
         }
-        "wg-mullvad".to_string()
+        MULLVAD_INTERFACE_NAME.to_string()
     }
 
     fn stop(mut self: Box<Self>) -> std::result::Result<(), TunnelError> {
@@ -393,26 +386,7 @@ impl Tunnel for NetworkManager {
                 log::error!("Failed to fetch WireGuard device config: {}", err);
                 TunnelError::GetConfigError
             })?;
-
-            // iterate over device attributes
-            let mut tx_bytes = 0;
-            let mut rx_bytes = 0;
-            for nla in device.nlas {
-                if let DeviceNla::Peers(peers) = nla {
-                    // iterate over all peer attributes
-                    let peer_iter = peers.iter().map(|peer| peer.0.as_slice()).flatten();
-
-                    for peer_nla in peer_iter {
-                        match peer_nla {
-                            PeerNla::TxBytes(bytes) => tx_bytes += *bytes,
-                            PeerNla::RxBytes(bytes) => rx_bytes += *bytes,
-                            _ => continue,
-                        };
-                    }
-                }
-            }
-
-            Ok(Stats { tx_bytes, rx_bytes })
+            Ok(Stats::parse_device_message(&device))
         });
 
         result
